+++ lib/cldr/backend/locale.ex
@@ -89,48 +89,6 @@ defmodule Cldr.Locale.Backend do
           Cldr.Locale.timezone_from_locale(locale)
         end
 
-        @doc """
-        Returns the localised display names data
-        for a locale name.
-
-        ## Arguments
-
-        * `locale` is any language tag returned by
-          `#{inspect(__MODULE__)}.new/1`
-          or a locale name in the list returned by
-          `#{inspect(config.backend)}.known_locale_names/0`
-
-        ## Returns
-
-        * A map of locale display names
-
-        ## Examples
-
-            => #{inspect(__MODULE__)}.display_names("en")
-
-        """
-        @doc since: "2.23.0"
-
-        @spec display_names(Cldr.LanguageTag.t() | Cldr.Locale.locale_name()) ::
-                {:ok, map()} | {:error, {module(), String.t()}}
-
-        def display_names(locale)
-
-        for locale_name <- Cldr.Config.known_locale_names(config) do
-          locale_display_names = Cldr.Config.get_locale(locale_name, config).locale_display_names
-
-          def display_names(unquote(locale_name)) do
-            {:ok, unquote(Macro.escape(locale_display_names))}
-          end
-        end
-
-        def display_names(%LanguageTag{} = locale) do
-          display_names(locale.cldr_locale_name)
-        end
-
-        def display_names(locale) do
-          {:error, Cldr.Locale.locale_error(locale)}
-        end
       end
     end
   end