+++ CHANGELOG.md
@@ -22,6 +22,8 @@ This is the changelog for Cldr v2.23.0 released on ______, 2021.  For older chan
 
 ### Enhancements
 
+* Add `Cldr.DisplayName` protocol definition to return a localised string representation of CLDR-based structs such as `t:Cldr.LanguageTag`, `t:Cldr.Unit` and `t:Cldr.Currency`
+
 * `Cldr.Locale.new/1,2` now passes all ~1600 validation tests for parsing and forming the canonical locale name. This is a prerequsite to impementing the [Locale Display Algorithm](https://unicode-org.github.io/cldr/ldml/tr35-general.html#Display_Name_Elements).
 
 * `Cldr.locale_and_backend_from/1` now supports a `map` of options as the argument.
@@ -36,7 +38,7 @@ This is the changelog for Cldr v2.23.0 released on ______, 2021.  For older chan
 
 * Add `Cldr.validate_script/1` to normalize and validate a script code (which is now in atom format as its canonical form)
 
-* Pre-compiled language tags (which are stored in `priv/cldr/language_tags.ebin`) are now cached during compilation resulting in a modest performance improvement in compile times.
+* Pre-compiled language tags (which are stored in `priv/cldr/language_tags.ebin`) are now cached during compilation resulting in a minor performance improvement in compile times.
 
 ## Cldr v2.22.1